requirements.txt
# Django core
Django==4.2.11

# REST API & Auth
djangorestframework==3.15.1
django-cors-headers==4.3.1
social-auth-app-django==5.4.1

# Async Tasks
celery==5.3.6
billiard==4.2.0
kombu==5.3.5
amqp==5.2.0

# Payments
django-paystack==1.0.0

# Database
psycopg2-binary==2.9.9

# UI
django-bootstrap5==23.4
django-crispy-forms==2.1

# Extra features
django-star-ratings==0.9.2
django-hitcount==1.3.5
django-model-utils==4.5.0
sorl-thumbnail==12.9.0
django-ckeditor==6.7.3
django-bootstrap-datepicker-plus==5.0.6
geopy==2.4.1

# Utilities
python-dateutil==2.9.0
pytz==2024.1
requests==2.32.3
simplejson==3.19.2
Pillow==10.4.0
six==1.17.0
openpyxl==3.1.5

# Environment
python-dotenv==1.0.1
django-environ==0.11.2

djangorestframework-simplejwt==5.3.1
django-filter==24.3
drf-spectacular==0.27.2